requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
claude-api==1.0.17
comm==0.2.1
curl-cffi==0.5.10
dataclasses-json==0.6.4
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.16.1
google-auth==2.27.0
google-generativeai==0.3.2
google-search-results==2.4.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
install==1.3.5
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langchain-google-genai==0.0.6
langchainhub==0.1.14
langsmith==0.0.86
lxml==5.1.0
marshmallow==3.20.2
matplotlib-inline==0.1.6
minify_html==0.15.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.3
openai==1.11.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
playwright==1.41.1
pluggy==1.4.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pyee==11.0.1
Pygments==2.17.2
pytest==7.4.4
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.2
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.13
wikipedia==1.4.0
yarl==1.9.4
